support acting as multiple slaves on the server
this is a breaking change, since the API to set up server slaves
is a bit different. also took the opportunity to change the default
slave id for a TCPServer

i.e. you may be a gateway device, acting on behalf of multiple slaves

also fix broadcast handling - don't expect or send responses to
broadcast commands