rework server message recognition and promiscuous mode
can't just rely on seeing a valid CRC, because they happen often _inside_
a longer valid message. so find a CRC, and then parse to make sure the
structure of the message is sane too. and then keep track of request/response
for promiscuous mode at that level, and don't force Server to reparse